pyproject.toml
[project]
name = "piframe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "numpy>=2.4.0",
    "piexif>=1.1.3",
    "pillow>=12.0.0",
    "requests>=2.32.5",
    "scikit-image>=0.26.0",
    "uvicorn>=0.40.0",
]
